@@ -2,7 +2,7 @@
 
 /**
  * @module lib/append
- * @copyright 2014- commenthol
+ * @copyright 2014-present commenthol
  * @license MIT
  */
 
@@ -19,8 +19,8 @@ function append (buffer, data, encoding) {
     return buffer
   }
   return Buffer.concat([
-    buffer || new Buffer(''),
-    (data instanceof Buffer ? data : new Buffer(data.toString(encoding || 'utf8')))
+    buffer || Buffer.from(''),
+    (data instanceof Buffer ? data : Buffer.from(data.toString(encoding || 'utf8')))
   ])
 }
 
