@@ -2,12 +2,12 @@
 
 /**
  * @module lib/Request
- * @copyright 2014- commenthol
+ * @copyright 2014-present commenthol
  * @license MIT
  */
 
 var util = require('util')
-var Readable = require('streamss-shim').Readable
+var Readable = require('stream').Readable
 var assign = require('./assign')
 var pick = require('mergee').pick
 
@@ -63,17 +63,17 @@ function Request (options) {
       rawHeaders: [],
       /** @see http://nodejs.org/api/http.html#http_message_trailers */
       trailers: {},
-      socket: {},                 // / {Object}  Dummy object for socket
-      connection: {               // / {Object}  Dummy object for connection
+      socket: {}, // / {Object}  Dummy object for socket
+      connection: { // / {Object}  Dummy object for connection
         remoteAddress: '127.0.0.1',
         remotePort: 51501
       },
       _internal: {
         httpVersion: '1.0',
-        buffer: new Buffer(''),
-        timer: 0,               // / {Number}  The time in msecs to take to
-        timedout: false,        // / {Boolean} If true than `Response.setTimeout` was called.
-        emitCloseCount: 0       // / {Number}
+        buffer: Buffer.from(''),
+        timer: 0, // / {Number}  The time in msecs to take to
+        timedout: false, // / {Boolean} If true than `Response.setTimeout` was called.
+        emitCloseCount: 0 // / {Number}
       }
     },
     pick(options, 'url,method,trailers')
@@ -131,7 +131,7 @@ Request.prototype._read = function (size) {
     this._internal.buffer = this._internal.buffer.slice(size, this._internal.buffer.length)
     this._internal.emitCloseCount += size
     if (this._internal.emitClose && this._internal.emitClose <= this._internal.emitCloseCount) {
-      this._internal.buffer = new Buffer('') // reset buffer to stop pushing
+      this._internal.buffer = Buffer.from('') // reset buffer to stop pushing
       this.emit('close')
     }
     if (!this.push(buf)) {
@@ -166,7 +166,7 @@ assign(Request.prototype, {
    * @param {String} [encoding]
    */
   setBuffer: function (data, encoding) {
-    this._internal.buffer = new Buffer(data.toString(encoding || 'utf8'))
+    this._internal.buffer = Buffer.from(data.toString(encoding || 'utf8'))
   },
   /**
    * Test only - Set HTTP header
