@@ -2,12 +2,12 @@
 
 /**
  * @module lib/Response
- * @copyright 2014- commenthol
+ * @copyright 2014-present commenthol
  * @license MIT
  */
 
 var util = require('util')
-var Writable = require('streamss-shim').Writable
+var Writable = require('stream').Writable
 var assign = require('./assign')
 var append = require('./append')
 var pick = require('mergee').pick
@@ -52,20 +52,20 @@ function Response (options) {
 
   assign(this, {
     _internal: {
-      headers: {},                // / {Object}  Response headers
-      rawHeaders: {},             // / {Object}  Raw Response headers
-      trailers: {},               // / {Object}  Trailing Response headers
-      buffer: new Buffer(''),     // / {Buffer}  Internal buffer represents response body
-      timedout: false,            // / {Boolean} If true than `Response.setTimeout` was called.
-      ended: false,               // / {Boolean} If true than `Response.end` was called.
-      timer: null                // / {Timer}   The timer object used by `Response.setTimeout`
+      headers: {}, // / {Object}  Response headers
+      rawHeaders: {}, // / {Object}  Raw Response headers
+      trailers: {}, // / {Object}  Trailing Response headers
+      buffer: Buffer.from(''), // / {Buffer}  Internal buffer represents response body
+      timedout: false, // / {Boolean} If true than `Response.setTimeout` was called.
+      ended: false, // / {Boolean} If true than `Response.end` was called.
+      timer: null // / {Timer}   The timer object used by `Response.setTimeout`
     },
-    socket: {},                     // / {Object}  Dummy object for socket
-    connection: {},                 // / {Object}  Dummy object for connection
-    statusCode: undefined,          /** @see http://nodejs.org/api/http.html#http_response_statuscode */
-    statusMessage: '',              /** @see http://nodejs.org/api/http.html#http_response_statusmessage */
-    headersSent: false,             /** @see http://nodejs.org/api/http.html#http_response_headerssent */
-    sendDate: true                 /** @see http://nodejs.org/api/http.html#http_response_senddate */
+    socket: {}, // / {Object}  Dummy object for socket
+    connection: {}, // / {Object}  Dummy object for connection
+    statusCode: undefined, /** @see http://nodejs.org/api/http.html#http_response_statuscode */
+    statusMessage: '', /** @see http://nodejs.org/api/http.html#http_response_statusmessage */
+    headersSent: false, /** @see http://nodejs.org/api/http.html#http_response_headerssent */
+    sendDate: true /** @see http://nodejs.org/api/http.html#http_response_senddate */
   }
   )
 
