@@ -2,7 +2,7 @@
 
 /**
  * @module lib/assign
- * @copyright 2014- commenthol
+ * @copyright 2014-present commenthol
  * @license MIT
  */
 
